Review working and skill search too

+++ Knowlify.Client/src/services/review.ts
@@ -3,46 +3,60 @@ import { IReview } from '@/interfaces/review'
 
 export const getReviews = async () => {
   const response = await fetch(API_ROUTES.review.all)
-  const data = (await response.json()) as IReview[]
-  return data
+  return response
 }
 
 export const getReview = async (id: number) => {
   const response = await fetch(`${API_ROUTES.review.get}/${id}`)
-  const data = (await response.json()) as IReview
-  return data
+  return response
 }
 
-export const addReview = async (review: IReview) => {
+export const getReviewsByBarter = async (barterId: number) => {
+  const response = await fetch(`${API_ROUTES.review.ByBarterId}/${barterId}`)
+  return response
+}
+
+export const addReview = async (review: IReview, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.review.add, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(review),
   })
-  const data = (await response.json()) as IReview
-  return data
+  return response
 }
 
-export const updateReview = async (review: IReview) => {
+export const updateReview = async (review: IReview, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(API_ROUTES.review.update, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
+      'Authorization': `Bearer ${token}`,
     },
     body: JSON.stringify(review),
   })
-  const data = (await response.json()) as IReview
-  return data
+  return response
 }
 
-export const deleteReview = async (id: number) => {
+export const deleteReview = async (id: number, token: string) => {
+  if (token === '') {
+    throw new Error('No token provided')
+  }
   const response = await fetch(`${API_ROUTES.review.delete}/${id}`, {
     method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${token}`,
+    },
   })
-  const data = (await response.json()) as IReview
-  return data
+  return response
 }
 
 // export const getReviewsByTransaction = async (id: number) => {
